+++ src/com/ft/jenkins/slack/SlackUtils.groovy
